Remove InfinispanMultiSiteLoadBalancerCheckProviderFactory.java
