+++ model/infinispan/src/main/resources/META-INF/services/org.keycloak.health.LoadBalancerCheckProviderFactory
@@ -1,2 +1,1 @@
-org.keycloak.connections.infinispan.InfinispanMultiSiteLoadBalancerCheckProviderFactory
 org.keycloak.connections.infinispan.remote.RemoteLoadBalancerCheckProviderFactory
